Move back to test_unicode; open-code loop; mark as uses-CPU.
